Рефакторинг главного экрана (Home)
