+++ src/blocks/home-feed/index.js
@@ -1,0 +1,1 @@
+export { default } from './home-feed';
