+++ src/blocks/home-top-bar/index.js
@@ -1,0 +1,1 @@
+export { default } from './home-top-bar';
