+++ src/blocks/home/__header/home__header.scss
@@ -1,0 +1,7 @@
+.home__header {
+    position: fixed;
+    width: 100%;
+    background-color: #ffffff;
+    z-index: 999;
+    box-shadow: 0 1px 4px 0 rgba(0, 0, 0, 0.2);
+}
