--- src/blocks/home/home.scss
@@ -1,9 +1,0 @@
-.home {
-    &__floating-header {
-        position: fixed;
-        width: 100%;
-        background-color: #ffffff;
-        z-index: 999;
-        box-shadow: 0 1px 4px 0 rgba(0, 0, 0, 0.2);
-    }
-}
